CRUD playlist listo

+++ src/main/java/parcial/backend/repositories/PlaylistRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import parcial.backend.entities.Playlist;
 @Repository
-public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
+public interface PlaylistRepository extends JpaRepository<Playlist, Integer> {
 }
